Nearly working version of dialop rollout generator
Some of the dialop env rules are still wrong

note generate_realistic_example.py will generate the tokens/masks that
will be SFT'd on. the reward should be the normalized reward.

heavily clauded. some design choices may not make sense, but plan is
quite thorough and accounts for many of the possible pitfalls.

this design was deemed superior to both:
- integrating with verl more, implementing new Interaction and
  RewardManager to work with existing sglang rollout code (this would
  suffer from having to do a ton of hacks to make things work well with
  the constraints they impose)
- integrating with verl less, using dialop to generate sequences and
  then using verl preexisting code to SFT on the sequences (this would
  not save us any effort in trying to get the dataset to adhere to a
  verl-compatible format, plus we would have to reimplement the outer
  GRPO loop and some of the core GRPO reward estimation logic, and we
  would miss out on verl's optimization)