At this point the rollout generator is working correctly
We get successful runs and the mask behavior looks correct, as well as
the observation of errors.

Masks may be off by 1 or 2 on either side, I think technically you
shouldn't train on "start of assistant sequence" tokens. But that's
okay. I just want to see if this basically works first.