Fix the GRPO grouping strategy
you should group with things of the "same" setting, so group by game.

not sure if this really matters, because the games are so different. but
try it anyway.

i used to avoid doing this because generation was so slow. but because
the sglang parallelism is working, it's not nearly as bad.