revert parameter offloading
the orchestrating script was assigning the gpus wrong, that's why i
didn't have as much memory as i used to and so this original way works.

fsdp sharding instead of fsdp2 nd offloading params is 10 times slower.
gpt5/claude4 told me to expect it to be about 2 times slower, so i
thought it wouldn't be that big of a deal and it could be made up by
batch size.

also this param offloading for some reason make computing the validation
loss EXTREMELY slow. probably because the params have to be set to
eval() or something.

i don't know how the implementation works but it doesn't save anywhere
near as much memory as you would think, for some reason, so even that
doesn't work, and we should just stick to fsdp2 now that we know that it
works