fix the thing where sft only works on 2 gpus
gpt-5 found this with a good prompt. apparently there is some setting in
NCCL that tries some kind of P2P operation and this just will not work
(hangs) because some hardware level thing is missing if the GPUs are not
NVLinked. you can just set NCCL to not try this if the GPUs aren't
NVLinked. why this isn't automatically detected by NCCL i'm not sure.

unfortunately computation isn't any faster, maybe due to the
communication overhead, but it frees enough memory that we can triple
the batch size.

at this point maybe we can try to optimze the memory usage of the sglang
thing