Implement some complicated resampling logic
log the reward better during the rollout phase, and compute stats

basically if in GRPO the thing doesn't keep improving, just destroy
progress and resample. i could not find an obvious pattern in variance
in the training data vs performance in the next round. (of course worse
performance predicted higher variance in the next round.)