Support filter by percentile; saving rollouts mid-generation
The percentile is for faster convergence on GRPO, but it seems to cause
an overly narrow policy.
The saving rollouts behavior is for mass-generating rollouts on one
policy. I think this probably isn't useful because it'll lead to policy
collapse as well.

Also collect mean reward in wandb and mean length stat.