Actually fix the mask and logprobs alignment issues now
this is more complex than previously appreciated, and also the previous
logic was completely wrong

basically qwen3's chat template removes previous thinking blocks from
the context. we're stuck between two bad places  with this because
1. if we go with it, then training as one sequence won't provide real
   feedback on the behavior in early turns; if anything it incentivizes
   the model to not think
2. if we don't go with it, and we manually insert the thinking blocks
   back into context, then we have moved the model away from its
   pretraining distribution, so its performance may weaken.

the "correct" thing to do is to follow the chat template (thereby
staying in distrubtion), and then treat every response by the model as
its own training example (thereby giving feedback on every think block).
this is the approach suggested by the qwen maintainers, e.g, in
https://github.com/QwenLM/Qwen3/discussions/1398 . i did not do that
because it would be extremely complicated and also possibly kind of
slow.

note that verl/ also has this issue because it also relies on the chat
template to provide one training example, both in the SFT and RL
workflows. slime doesn't have the exact same issue, since it
reconstructs the sequence directly, but it still treats one chat
sequence as one training example, so it has the same behavior as our
workaround/compromise.