update the training script for multiple cases
1. put all the args in a list so they can be commented without breaking
   bash parsing

2. add some args for the wiki eval, for the initial sft in particular

3. modify args for new liger kernel setup + bugfixes

some insights from recently:

the p99 of the gpt-5 conversations is about 4600 tokens
activation checkpointing is necessary. forward pass for a single example
is about 16gb, so if you don't do this your batch size will never be
above 2 even with all the special kernels

lora allows a batch size of 16 per gpu, but it's not that much faster
because time per iteration increases anyway, i guess due to the
increased compute requirement from the checkpointing? different
checkpointing strategy may be better but i don't want to mess with that

one way of thinking of this is i already traded memory for compute when
i enabled activation checkpointing.

last possible thing is lora + no activation checkpointing + smaller
batch size could be a lot faster? iterations could be as fast as 7 or 8s
and now iterations take 45s or so